requirements.txt
Flask
Flask-WTF
SQLAlchemy
argon2-cffi
Flask-Limiter
Werkzeug

